changed type
git-svn-id: svn://orca.physics.unc.edu/Orca@1855 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Main Dialogs/ORCARootService/ORCARootServiceController.m
@@ -268,7 +268,7 @@ SYNTHESIZE_SINGLETON_FOR_CLASS(ORCARootServiceController);
 	return  [[self orcaRootService] connectionHistoryCount];
 }
 
-- (id)comboBox:(NSComboBox *)aComboBox objectValueForItemAtIndex:(NSInteger)index
+- (id)comboBox:(NSComboBox *)aComboBox objectValueForItemAtIndex:(unsigned)index
 {
 	return [[self orcaRootService] connectionHistoryItem:index];
 }
